+++ src/input_parsers/OnnxParser.h
@@ -94,6 +94,7 @@ private:
     void scaleAndAddEquations( onnx::NodeProto& node, bool makeEquations, double coefficient1, double coefficient2 );
     void matMulEquations( onnx::NodeProto& node, bool makeEquations );
     void reluEquations( onnx::NodeProto& node, bool makeEquations );
+    void leakyReluEquations( onnx::NodeProto& node, bool makeEquations );
     void sigmoidEquations( onnx::NodeProto& node, bool makeEquations );
     void tanhEquations( onnx::NodeProto& node, bool makeEquations );
 };