+++ CHANGELOG.md
@@ -11,7 +11,7 @@
 * Added support for Leaky ReLU
 * Added support for Pythonic API
 * Command-line ONNX parser now supports networks with multiple outputs.
-* Command-line ONNX parser now supports the following operators: Cast, Squeeze.
+* Command-line ONNX parser now supports the following operators: Cast, Squeeze, LeakyRelu.
 * Errors now are printed on `stderr` rather than `stdout`
 * Not reindexing output variables to immediately follow after input variables in the MarabouNetworkONNX class